backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.32
psycopg[binary]>=3.2.2
alembic==1.13.2
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.2
fastapi-admin==1.0.4
email-validator==2.2.0
setuptools>=65.0.0

